src/main/java/uk/gov/digital/ho/hocs/info/api/dto/ParentTopicDto.java
package uk.gov.digital.ho.hocs.info.api.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Getter;
import uk.gov.digital.ho.hocs.info.domain.model.ParentTopic;

import java.util.UUID;

@AllArgsConstructor(access = AccessLevel.PRIVATE)
@Getter
public class ParentTopicDto {

    @JsonProperty("label")
    private String displayName;

    @JsonProperty("value")
    private UUID uuid;

    public static ParentTopicDto from (ParentTopic parentTopic) {
        return new ParentTopicDto(parentTopic.getDisplayName(), parentTopic.getUuid()); }
}
